backend/ai/app/modules/transcription/requirements.txt
websockets
websocket-client
pyaudio
python-dotenv
numpy
transformers
torch
